Make MapableConfigValues extend ConfigValue<T>
